Edit : Main.scss FeedTop component 수정
